+++ functorch/_src/aot_autograd/__init__.py
@@ -1,0 +1,7 @@
+# This file has moved. It is not public API. If you are not a PyTorch developer
+# and you are relying on the following imports, please file an issue.
+from torch._functorch.aot_autograd import (
+    aot_autograd_decompositions,
+    KNOWN_TYPES,
+    PytreeThunk,
+)
