+++ functorch/_src/eager_transforms/__init__.py
@@ -1,0 +1,6 @@
+# This file has moved. It is not public API. If you are not a PyTorch developer
+# and you are relying on the following imports, please file an issue.
+from torch._functorch.eager_transforms import (
+    _unwrap_functional_tensor,
+    _assert_wrapped_functional,
+)
