+++ functorch/_src/make_functional/__init__.py
@@ -1,0 +1,2 @@
+# This file has moved. Please update your imports
+from torch._functorch.make_functional import _swap_state
