+++ test/dynamo/test_aot_cudagraphs.py
@@ -104,7 +104,7 @@ class TestAotCudagraphs(torch._dynamo.test_case.TestCase):
         y = torch.randn((), device="cpu")
         fn(x, y)
 
-    @patch("functorch._src.config.use_functionalize", True)
+    @patch("torch._functorch.config.use_functionalize", True)
     def test_mutate_input(self):
         def model(x, y):
             y.add_(3)
@@ -159,7 +159,7 @@ class TestAotCudagraphs(torch._dynamo.test_case.TestCase):
         y = torch.randn(3, device="cuda:0", requires_grad=True)
         fn(y)
 
-    @patch("functorch._src.config.use_functionalize", True)
+    @patch("torch._functorch.config.use_functionalize", True)
     @patch_all()
     def test_mutated_metadata(self):
         # more tortured example at
@@ -180,7 +180,7 @@ class TestAotCudagraphs(torch._dynamo.test_case.TestCase):
         x = torch.empty(0, device="cuda:0")
         fn(x)
 
-    @patch("functorch._src.config.use_functionalize", True)
+    @patch("torch._functorch.config.use_functionalize", True)
     @patch_all()
     def test_dead_fill(self):
         def model(x):
