+++ test/dynamo/test_repros.py
@@ -11,14 +11,14 @@ from copy import deepcopy
 from typing import List
 from unittest.mock import patch
 
-import functorch._src.config
-
 import numpy as np
 import torch
 
 import torch._dynamo.test_case
 import torch._dynamo.testing
 import torch._dynamo.utils
+
+import torch._functorch.config
 
 try:
     from test_minifier import requires_cuda
@@ -1681,7 +1681,7 @@ class ReproTests(torch._dynamo.test_case.TestCase):
         opt_fn(x)
         self.assertEqual(cnt.frame_count, 1)
 
-    @patch.object(functorch._src.config, "use_dynamic_shapes", True)
+    @patch.object(torch._functorch.config, "use_dynamic_shapes", True)
     def test_bigbird_unsqueeze_inplace(self):
         def fn(reshape_2):
             view_2 = reshape_2.clone()
@@ -2071,7 +2071,7 @@ class ReproTests(torch._dynamo.test_case.TestCase):
         with self.assertRaisesRegex(torch._dynamo.exc.Unsupported, "generic_jump"):
             torch._dynamo.export(f, torch.Tensor([3, 4, 5]))
 
-    @patch.object(functorch._src.config, "use_dynamic_shapes", True)
+    @patch.object(torch._functorch.config, "use_dynamic_shapes", True)
     def test_batchnorm_e2e(self):
         class Repro(torch.nn.Module):
             def __init__(self):
