+++ test/test_functionalization.py
@@ -157,7 +157,7 @@ class TestFunctionalization(TestCase):
 
         def g(x):
             loss = f(x).sum()
-            from functorch._src.aot_autograd import setup_stacktrace_preservation_hooks
+            from torch._functorch.aot_autograd import setup_stacktrace_preservation_hooks
             import torch.fx.traceback as fx_traceback
             setup_stacktrace_preservation_hooks([loss.grad_fn])
             with fx_traceback.override_stack_trace():
